requirements.txt
streamlit
langgraph
supabase
openai
python-dotenv
pandas
requests
lark
fastapi
uvicorn
open-autonomy[all]
open-aea-ledger-ethereum 